Modify the  new enterpriseSetting
